setup all ui elements

+++ app/src/main/java/kozlov/artyom/avitoweather/data/network/pojo/Temp.kt
@@ -2,7 +2,7 @@ package kozlov.artyom.avitoweather.data.network.pojo
 
 data class Temp(
     val day: Double,
-    val eve: Int,
+    val eve: Double,
     val max: Double,
     val min: Double,
     val morn: Double,
